Porting on 2.1.1.
Starting with the OpenFOAM-2.1.1, the dimensioned<Type> interface has  the default contructor.
Some other interfaces have functions which return dimensioned<Type>(for example GeometricField:weightedAverage). When these functions are being wrapped, we receive generated code:
.........
dimensioned<Type> result;
.....................................
result = GeometricField-> weightedAverage( arg ).
..............
This code is wrong because the dimension of the left operand is [0,0,0,0](see default constructor ) and the dimension of the right is  other. So we have wrong code.
We add feature "valuewrapper" which transforms line "dimensioned<Type> result;"  to "SwigValueWrapper<dimensioned<Type>>"